+++ src/Data/Profunctor/Composition.hs
@@ -284,9 +284,9 @@ instance p ~ q => Category (Rift p q) where
   Rift f . Rift g = Rift (g . f)
   {-# INLINE (.) #-}
 
--- | The 2-morphism that defines a left Kan lift.
+-- | The 2-morphism that defines a right Kan lift.
 --
--- Note: When @p@ is right adjoint to @'Rift' p (->)@ then 'decomposeRift' is the 'counit' of the adjunction.
+-- Note: When @p@ is left adjoint to @'Rift' p (->)@ then 'decomposeRift' is the 'counit' of the adjunction.
 decomposeRift :: Procompose p (Rift p q) :-> q
 decomposeRift (Procompose p (Rift pq)) = pq p
 {-# INLINE decomposeRift #-}